Finish MultiSupplier

+++ CuentasClaras/Api/Organisation/ReleaseItemDTO.cs
@@ -15,7 +15,7 @@ namespace CuentasClaras.Model
         public string UnitName { get; set; }
         public string Description { get; set; }
         public int Quantity { get; set; }
-        public int TotalAmountUYU { get; set; }
+        public double TotalAmountUYU { get; set; }
         //FK
         public int? ClassificationId { get; set; }
         public string Classification { get; set; }
